requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-google-genai>=0.0.3
langchain-huggingface>=0.0.3
chromadb>=0.4.22
pandas>=2.0.0
langchain-chroma
tenacity>=8.2.3
python-dotenv>=1.0.0
pytest>=8.0.0
tiktoken
sentence-transformers